api_yamdb/requirements.txt
requests==2.26.0
django==2.2.16
djangorestframework==3.12.4
django-filter
djangorestframework-simplejwt
PyJWT==2.1.0
drf-yasg
gunicorn
psycopg2-binary
python-dotenv
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
